Move SaveChannelUseCase to domain module
