--- app/src/main/java/dev/punitd/feedly/onboarding/domain/usecase/SaveChannelUseCase.kt
@@ -1,7 +1,0 @@
-package dev.punitd.feedly.onboarding.domain.usecase
-
-import dev.punitd.data.Channel
-
-interface SaveChannelUseCase {
-    suspend fun execute(channel: Channel): Unit
-}
