+++ app/src/main/java/dev/punitd/feedly/onboarding/inject/OnboardingModule.kt
@@ -6,8 +6,6 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
 import dev.punitd.feedly.onboarding.domain.usecase.GetPrefilledChannelsUseCase
 import dev.punitd.feedly.onboarding.domain.usecase.GetPrefilledChannelsUseCaseImpl
-import dev.punitd.feedly.onboarding.domain.usecase.SaveChannelUseCase
-import dev.punitd.feedly.onboarding.domain.usecase.SaveChannelUseCaseImpl
 
 @InstallIn(ViewModelComponent::class)
 @Module
@@ -17,9 +15,4 @@ abstract class OnboardingModule {
     abstract fun provideGetPrefilledChannelsUseCase(
         getPrefilledChannelsUseCase: GetPrefilledChannelsUseCaseImpl
     ): GetPrefilledChannelsUseCase
-
-    @Binds
-    abstract fun provide(
-        saveChannelUseCase: SaveChannelUseCaseImpl
-    ): SaveChannelUseCase
 }
