+++ core/domain/src/main/java/dev/punitd/domain/inject/DomainModule.kt
@@ -8,6 +8,8 @@ import dev.punitd.domain.repository.ChannelsRepository
 import dev.punitd.domain.repository.ChannelsRepositoryImpl
 import dev.punitd.domain.usecase.GetSavedChannelsUseCase
 import dev.punitd.domain.usecase.GetSavedChannelsUseCaseImpl
+import dev.punitd.domain.usecase.SaveChannelUseCase
+import dev.punitd.domain.usecase.SaveChannelUseCaseImpl
 
 @InstallIn(ViewModelComponent::class)
 @Module
@@ -17,6 +19,11 @@ abstract class DomainModule {
     abstract fun provideGetSavedChannelsUseCase(
         getSavedChannelsUseCase: GetSavedChannelsUseCaseImpl
     ): GetSavedChannelsUseCase
+
+    @Binds
+    abstract fun provideSaveChannelUseCase(
+        saveChannelUseCase: SaveChannelUseCaseImpl
+    ): SaveChannelUseCase
 
     @Binds
     abstract fun provideChannelsRepo(
